test/test.js simply require()s other test suites
